+++ app/src/main/java/com/example/ankiscan/data/DictDao.kt
@@ -10,17 +10,6 @@ import kotlinx.coroutines.flow.Flow
 
 @Dao
 interface DictDao {
-//    @Insert(onConflict = OnConflictStrategy.ABORT)
-//    suspend fun insert(dictEntry: DictEntry)
-//
-//    @Update
-//    suspend fun update(dictEntry: DictEntry)
-//
-//    @Delete
-//    suspend fun delete(dictEntry: DictEntry)
-//
-//    @Query("SELECT * from dictEntries WHERE id = :id")
-//    fun getEntry(id: Int): Flow<DictEntry>
     @Insert(onConflict = OnConflictStrategy.ABORT)
     suspend fun insertReadingElement(readingElement: ReadingElement)
 
@@ -30,4 +19,18 @@ interface DictDao {
     @Insert(onConflict = OnConflictStrategy.ABORT)
     suspend fun insertSense(sense: Sense)
 
+    @Query("SELECT entry_id from reading_element WHERE r_element = :element")
+    fun getReadingElementEntryNumber(element: String): Flow<Int>
+
+    @Query("SELECT entry_id from kanji_element WHERE k_element = :element")
+    fun getKanjiElementEntryNumber(element: String): Flow<Int>
+
+    @Query("SELECT * from reading_element WHERE entry_id = :entryId")
+    fun getReadingElement(entryId: Int): Flow<ReadingElement>
+
+    @Query("SELECT * from kanji_element WHERE entry_id = :entryId")
+    fun getKanjiElement(entryId: Int): Flow<KanjiElement>
+
+    @Query("SELECT * from gloss WHERE entry_id = :entryId")
+    fun getSense(entryId: Int): Flow<Sense>
 }
